fixed map view not displaying when resuming activity. sped up table display UI.
Fragments in TableDisplayActivity are now re-used, so that they are only ever
initialized once per viewing of a table (excluding GraphView and DetailView).
No tests were introduced for this, to avoid getting locked in by the tests, but
if things start to misbehave tests should be created for the changes
introduced at this phase.

--HG--
branch : development